+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.0.3]
 
 ### Fixes
+- [Issue #3] Some errors are not logged with the error string
 
 ## [1.0.2]