+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # ts-tiny-log
 
+## [1.0.3]
+
+### Fixes
+- [Issue #3] Some errors are not logged with the error string
+
 ## [1.0.2]
 
 ### Fixes